requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.1.18
stripe==7.8.0
google-cloud-secret-manager==2.18.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
gunicorn==21.2.0
python-decouple==3.8
django-extensions==3.2.3
dj-database-url==2.1.0
drf-spectacular==0.26.5
django-ratelimit==4.1.0
django-storages==1.14.2
boto3==1.34.0

# A2A SDK and Agent Broker Dependencies
# google-cloud-aiplatform==1.36.0
google-cloud-discoveryengine==0.11.4
google-cloud-dialogflow==2.24.1
# google-cloud-vertexai==1.36.0

# Payment Processor Dependencies (AP2)
Adyen==13.0.0
plaid-python==9.0.0
requests==2.31.0
cryptography==41.0.7

# Note: Slack integration handled by Salesforce team via Agentforce

# Additional utilities
pydantic==2.5.0
# asyncio==3.4.3
